requirements_unified.txt
# OMNI Unified Platform Requirements
# Vse potrebne odvisnosti za združeno OMNI platform aplikacijo

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
gunicorn>=20.0.0

# HTTP and networking
httpx>=0.24.0
requests>=2.28.0
aiohttp>=3.8.0

# Google Cloud and Vertex AI
google-cloud-aiplatform>=1.30.0
google-cloud-storage>=2.8.0
google-cloud-pubsub>=2.15.0
google-cloud-tasks>=2.10.0
google-generativeai>=0.3.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Data processing and visualization
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.15.0

# System monitoring
psutil>=5.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.5
jinja2>=3.1.0

# Security (compatible versions)
cryptography>=40.0.0
bcrypt>=4.0.0

# JSON and data formats
pyyaml>=6.0.0

# Time and date utilities
python-dateutil>=2.8.0
pytz>=2023.3